+++ Tests/LibraryCore.Tests.Core/Parsers/RuleParser/Tokens/LambdaParserTest.cs
@@ -1,0 +1,51 @@
+﻿using LibraryCore.Core.Parsers.RuleParser.TokenFactories.Implementation;
+using LibraryCore.Tests.Core.Parsers.RuleParser.Fixtures;
+using System.Linq.Expressions;
+
+namespace LibraryCore.Tests.Core.Parsers.RuleParser.Tokens;
+
+public class LambdaParserTest : IClassFixture<RuleParserFixture>
+{
+    public LambdaParserTest(RuleParserFixture ruleParserFixture)
+    {
+        RuleParserFixture = ruleParserFixture;
+    }
+
+    private RuleParserFixture RuleParserFixture { get; }
+
+    [Fact]
+    public void ParseTest()
+    {
+        Assert.False(true);
+        //var result = RuleParserFixture.ResolveRuleParserEngine()
+        //                                     .ParseString("1 == 1")
+        //                                     .CompilationTokenResult;
+
+        //Assert.Equal(5, result.Count);
+        //Assert.IsType<NumberToken<int>>(result[0]);
+        //Assert.IsType<WhiteSpaceToken>(result[1]);
+        //Assert.IsType<EqualsToken>(result[2]);
+        //Assert.IsType<WhiteSpaceToken>(result[3]);
+        //Assert.IsType<NumberToken<int>>(result[4]);
+    }
+
+    //[Fact]
+    //public void CreateTokenNotImplemented() => Assert.Throws<NotImplementedException>(() => new EqualsToken().CreateExpression(Array.Empty<ParameterExpression>()));
+
+    //[InlineData("$Survey.PriceOfSurgery$ == 9.99d", true)]
+    //[InlineData("$Survey.PriceOfSurgery$ == 8.25d", false)]
+    //[InlineData("$Survey.PriceOfSurgery$ == 60d", false)]
+    //[InlineData("$Survey.Name$ == 'Jacob DeGrom'", true)]
+    //[InlineData("$Survey.Name$ == 'Tommy'", false)]
+    //[Theory]
+    //public void ExpressionsToTest(string expressionToTest, bool expectedResult)
+    //{
+    //    var expression = RuleParserFixture.ResolveRuleParserEngine()
+    //                                        .ParseString(expressionToTest)
+    //                                        .BuildExpression<Survey>("Survey")
+    //                                        .Compile();
+
+    //    Assert.Equal(expectedResult, expression.Invoke(new SurveyModelBuilder().Value));
+    //}
+}
+
